Move specific mixins logic from base to common
They aren't really base classes or "core" mixins but just providing
functionality for cases where we cannot reuse the same RH for both
manager and track reviewer views.
